SOURCE: fixed a bug caused by executing the source() command in an R script
The problem is that the source('scriptname.R') command loads the content
of the script in the same environment of the calling script.

What does this mean?

If there are some variables with the same name they will be overwritten!!

This was causing an issue with the subsequent instructions in the program.