QUANTILES.R: added method to compute the quantiles precisely
Method 1 relies on the cumulative frequency distribution.

Method 2 is more precise in that it computes the desired rank as
the product of the percentile and the number of observations.

Then if there is any fractional part we also take that into consideration.

For example, suppose to compute the 5% on 299 values.

0.05 x 299 = 14.95

We need to compute a value that is 0.95 of the way between the
14th and 15th observations in rank order.