DESEQ: inspected unique read frequency by length
Since we are looking at RNA-seq data coming from short non-coding
RNAs, these typically are not fragmented, and moreover, the library
is highly likely to present duplicates.

Why?

Because the small RNAs have a function that depends on their sequence.

They are able to target molecules based on their sequence complementarity.

This means that typically there are many copies of the same small non-coding RNA.

That's why we see a high level of duplication in our data.

The bulk of the reads are 21 nucleotides long; this suggest siRNAs are
highly expressed.

Other small RNAs are not so highly expressed.

This view can help us better understand our data and avoid wrong interpretation.